requirements.txt
aiohttp==0.18.4
cchardet==1.0.0
chardet==2.3.0
decorator==4.0.4
docopt==0.6.2
flake8==2.5.0
greenlet==0.4.9
gunicorn==19.3.0
httpie==0.9.2
ipython==4.0.0
ipython-genutils==0.1.0
jedi==0.9.0
mccabe==0.3.1
msgpack-python==0.4.6
neovim==0.0.38
path.py==8.1.2
pep8==1.5.7
pexpect==4.0.1
pickleshare==0.5
prompt-toolkit==0.54
ptpython==0.25
ptyprocess==0.5
pyflakes==1.0.0
Pygments==2.0.2
PyJWT==1.4.0
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wcwidth==0.1.5
